Install "pactl" for Windows
+ Small fixes.

+++ scripts/shell/shellvnc_commands.sh
@@ -7,6 +7,7 @@ shellvnc_required_before_imports "${BASH_SOURCE[0]}" || return "$?" 2> /dev/null
 . "../messages/shellvnc_print_info_increase_prefix.sh" || shellvnc_return_0_if_already_sourced || return "$?" 2> /dev/null || exit "$?"
 . "../messages/shellvnc_print_text.sh" || shellvnc_return_0_if_already_sourced || return "$?" 2> /dev/null || exit "$?"
 . "../messages/shellvnc_print_error.sh" || shellvnc_return_0_if_already_sourced || return "$?" 2> /dev/null || exit "$?"
+. "../messages/shellvnc_print_warning.sh" || shellvnc_return_0_if_already_sourced || return "$?" 2> /dev/null || exit "$?"
 . "../messages/shellvnc_print_success_decrease_prefix.sh" || shellvnc_return_0_if_already_sourced || return "$?" 2> /dev/null || exit "$?"
 shellvnc_required_after_imports "${BASH_SOURCE[0]}" || return "$?" 2> /dev/null || exit "$?"
 
@@ -77,101 +78,83 @@ shellvnc_commands() {
     # Right now I am testing only on Arch.
     # ========================================
 
-    # Even if we can by default consider that the package name is the same as the command name, we don't want to accidentally install wrong package.
-    # And because this solution will has defined number of commands, we can easily add new commands to the list and make it more reliable.
-    local is_implemented=0
-
+    # (Only for Arch): If the package need to be installed from AUR (via "yay")
     local is_aur=0
 
+    # (Only for Windows): If the package is an installer (1) or a portable executable (0)
+    local is_windows_installer=0
+
+    # Even if we can by default consider that the package name is the same as the command name, we don't want to accidentally install wrong package.
+    # And because this solution will has defined number of commands, we can easily add new commands to the list and make it more reliable.
     # Linux: Required packages for the command;
     # Windows: URL to download the executable for command.
-    local package_name_or_link="${command}"
+    local package_name_or_link=""
 
     if [ "${command}" = "vncviewer" ]; then
       if [ "${_SHELLVNC_CURRENT_OS_NAME}" = "${_SHELLVNC_OS_NAME_ARCH}" ]; then
         package_name_or_link="tigervnc"
-        is_implemented=1
       elif [ "${_SHELLVNC_CURRENT_OS_NAME}" = "${_SHELLVNC_OS_NAME_FEDORA}" ]; then
         package_name_or_link="tigervnc"
-        is_implemented=1
       elif [ "${_SHELLVNC_CURRENT_OS_NAME}" = "${_SHELLVNC_OS_NAME_DEBIAN}" ]; then
         package_name_or_link="tigervnc-viewer"
-        is_implemented=1
       elif [ "${_SHELLVNC_CURRENT_OS_NAME}" = "${_SHELLVNC_OS_NAME_WINDOWS}" ]; then
         package_name_or_link="https://sourceforge.net/projects/tigervnc/files/stable/${TIGERVNC_VERSION_FOR_WINDOWS}/vncviewer64-${TIGERVNC_VERSION_FOR_WINDOWS}.exe/download"
-        is_implemented=1
       fi
     elif [ "${command}" = "jq" ]; then
       if [ "${_SHELLVNC_CURRENT_OS_NAME}" = "${_SHELLVNC_OS_NAME_ARCH}" ]; then
         package_name_or_link="jq"
-        is_implemented=1
       elif [ "${_SHELLVNC_CURRENT_OS_NAME}" = "${_SHELLVNC_OS_NAME_FEDORA}" ]; then
         package_name_or_link="jq"
-        is_implemented=1
       elif [ "${_SHELLVNC_CURRENT_OS_NAME}" = "${_SHELLVNC_OS_NAME_DEBIAN}" ]; then
         package_name_or_link="jq"
-        is_implemented=1
       elif [ "${_SHELLVNC_CURRENT_OS_NAME}" = "${_SHELLVNC_OS_NAME_WINDOWS}" ]; then
         package_name_or_link="https://github.com/jqlang/jq/releases/latest/download/jq-win64.exe"
-        is_implemented=1
       fi
     elif [ "${command}" = "vncserver" ]; then
       if [ "${_SHELLVNC_CURRENT_OS_NAME}" = "${_SHELLVNC_OS_NAME_ARCH}" ]; then
         package_name_or_link="tigervnc"
-        is_implemented=1
       elif [ "${_SHELLVNC_CURRENT_OS_NAME}" = "${_SHELLVNC_OS_NAME_FEDORA}" ]; then
         package_name_or_link="tigervnc-server"
-        is_implemented=1
       elif [ "${_SHELLVNC_CURRENT_OS_NAME}" = "${_SHELLVNC_OS_NAME_DEBIAN}" ]; then
         package_name_or_link="tigervnc-standalone-server"
-        is_implemented=1
       fi
     elif [ "${command}" = "tput" ]; then
       if [ "${_SHELLVNC_CURRENT_OS_NAME}" = "${_SHELLVNC_OS_NAME_TERMUX}" ]; then
         package_name_or_link="ncurses-utils"
-        is_implemented=1
       elif [ "${_SHELLVNC_CURRENT_OS_NAME}" = "${_SHELLVNC_OS_NAME_ARCH}" ]; then
         package_name_or_link="ncurses"
-        is_implemented=1
       elif [ "${_SHELLVNC_CURRENT_OS_NAME}" = "${_SHELLVNC_OS_NAME_FEDORA}" ]; then
         package_name_or_link="ncurses"
-        is_implemented=1
       elif [ "${_SHELLVNC_CURRENT_OS_NAME}" = "${_SHELLVNC_OS_NAME_DEBIAN}" ]; then
         package_name_or_link="ncurses-bin"
-        is_implemented=1
       fi
     elif [ "${command}" = "ssh" ]; then
       if [ "${_SHELLVNC_CURRENT_OS_NAME}" = "${_SHELLVNC_OS_NAME_ARCH}" ]; then
         package_name_or_link="openssh"
-        is_implemented=1
       elif [ "${_SHELLVNC_CURRENT_OS_NAME}" = "${_SHELLVNC_OS_NAME_FEDORA}" ]; then
         package_name_or_link="openssh-clients"
-        is_implemented=1
       elif [ "${_SHELLVNC_CURRENT_OS_NAME}" = "${_SHELLVNC_OS_NAME_DEBIAN}" ]; then
         package_name_or_link="openssh-client"
-        is_implemented=1
       fi
     elif [ "${command}" = "pactl" ]; then
       if [ "${_SHELLVNC_CURRENT_OS_NAME}" = "${_SHELLVNC_OS_NAME_ARCH}" ]; then
         package_name_or_link="libpulse"
-        is_implemented=1
       elif [ "${_SHELLVNC_CURRENT_OS_NAME}" = "${_SHELLVNC_OS_NAME_FEDORA}" ]; then
         package_name_or_link="pulseaudio-utils"
-        is_implemented=1
       elif [ "${_SHELLVNC_CURRENT_OS_NAME}" = "${_SHELLVNC_OS_NAME_DEBIAN}" ]; then
         package_name_or_link="pulseaudio-utils"
-        is_implemented=1
+      elif [ "${_SHELLVNC_CURRENT_OS_NAME}" = "${_SHELLVNC_OS_NAME_WINDOWS}" ]; then
+        shellvnc_print_warning "When installer shows up, please remember to enable \"${c_highlight}Allow module loading${c_return}\"!" || return "$?"
+        package_name_or_link="https://github.com/pgaskin/pulseaudio-win32/releases/download/${PULSEAUDIO_VERSION_FOR_WINDOWS}/pasetup.exe"
+        is_windows_installer=1
       fi
     elif [ "${command}" = "i3" ]; then
       if [ "${_SHELLVNC_CURRENT_OS_NAME}" = "${_SHELLVNC_OS_NAME_ARCH}" ]; then
         package_name_or_link="i3-wm"
-        is_implemented=1
       elif [ "${_SHELLVNC_CURRENT_OS_NAME}" = "${_SHELLVNC_OS_NAME_FEDORA}" ]; then
         package_name_or_link="i3"
-        is_implemented=1
       elif [ "${_SHELLVNC_CURRENT_OS_NAME}" = "${_SHELLVNC_OS_NAME_DEBIAN}" ]; then
         package_name_or_link="i3-wm"
-        is_implemented=1
       fi
     else
       # Commands which have the same package name for all Linux distributions
@@ -180,20 +163,18 @@ shellvnc_commands() {
         || [ "${_SHELLVNC_CURRENT_OS_NAME}" = "${_SHELLVNC_OS_NAME_DEBIAN}" ]; then
         if [ "${command}" = "pstree" ]; then
           package_name_or_link="psmisc"
-          is_implemented=1
         else
           local __same_command
           for __same_command in which sed grep git ssh scp screen sshpass usbip openbox; do
             if [ "${command}" = "${__same_command}" ]; then
               package_name_or_link="${__same_command}"
-              is_implemented=1
             fi
           done
         fi
       fi
     fi
 
-    if [ "${is_implemented}" = "0" ]; then
+    if [ -z "${package_name_or_link}" ]; then
       shellvnc_print_error "Installing command \"${c_highlight}${command}${c_return}\" is not implemented for OS \"${c_highlight}${_SHELLVNC_CURRENT_OS_NAME}${c_return}\"!" || return "$?"
       return 1
     fi
@@ -212,13 +193,29 @@ shellvnc_commands() {
       link_path="/usr/bin/${command}" || return "$?"
 
       if [ "${action}" = "${SHELLVNC_COMMANDS_ACTION_INSTALL}" ]; then
-        # 1. Download executable file
-        # 2. Create symlink so checks if this command is installed will work now
-        command_to_execute="sudo curl --fail -L -o \"${executable_path}\" \"${package_name_or_link}\" && sudo ln -s \"${executable_path}\" \"${link_path}\"" || return "$?"
+        if [ "${is_windows_installer}" = "0" ]; then
+          # 1. Download executable file
+          # 2. Create symlink so checks if this command is installed will work now
+          command_to_execute="sudo curl --fail -L -o \"${executable_path}\" \"${package_name_or_link}\" && sudo ln -sf \"${executable_path}\" \"${link_path}\""
+        else
+          local installer_file_name
+          installer_file_name="$(basename "${executable_path}")" || return "$?"
+
+          # 1. Download installer
+          # 2. Run installer
+          # 3. Remove installer
+          command_to_execute="sudo curl --fail -L -o \"${installer_file_name}\" \"${package_name_or_link}\" && ./${installer_file_name} || { error_code=\"\$?\" && rm \"${installer_file_name}\"; shellvnc_print_error \"Error occurred while trying to download and run the installer!\" || return \"\$?\"; return \"\${error_code}\"; }; rm \"${installer_file_name}\"" || return "$?"
+        fi
       elif [ "${action}" = "${SHELLVNC_COMMANDS_ACTION_UNINSTALL}" ]; then
-        # 1. Delete executable file
-        # 2. Remove symlink
-        command_to_execute="sudo rm -rf \"${executable_path}\" && sudo unlink \"${link_path}\"" || return "$?"
+        if [ "${is_windows_installer}" = "0" ]; then
+          # 1. Delete executable file
+          # 2. Remove symlink
+          command_to_execute="sudo rm -rf \"${executable_path}\" && sudo unlink \"${link_path}\"" || return "$?"
+        else
+          # TODO: Implement uninstallation for Windows
+          shellvnc_print_error "Uninstalling \"${c_highlight}${command}${c_return}\" is not implemented for \"${c_highlight}${_SHELLVNC_CURRENT_OS_NAME}${c_return}\"!" || return "$?"
+          return 1
+        fi
       else
         shellvnc_print_error "Unknown action \"${c_highlight}${action}${c_return}\"!" || return "$?"
         return 1
